+++ src/Transformer/String/RemoveWhitespace/RemoveWhitespaceTransformer.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ExtendsSoftware\ExaPHP\Transformer\String\RemoveWhitespace;
+
+use ExtendsSoftware\ExaPHP\Transformer\TransformerInterface;
+
+use function is_string;
+use function preg_replace;
+
+class RemoveWhitespaceTransformer implements TransformerInterface
+{
+    /**
+     * @inheritDoc
+     */
+    public function transform(mixed $value): mixed
+    {
+        if (!is_string($value)) {
+            return $value;
+        }
+
+        return preg_replace('/\s+/u', '', $value);
+    }
+}